pull out ecr summary to its own file
